+++ packages/heroine/lib/src/heroine_velocity.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:superhero/src/drag_dismissable.dart';
-import 'package:superhero/src/superheroes.dart';
+import 'package:heroine/src/drag_dismissable.dart';
+import 'package:heroine/src/heroines.dart';
 
 /// An inherited widget that can be used to provide a velocity to the next
-/// [Superhero] below this widget.
+/// [Heroine] below this widget.
 ///
-/// When a [Superhero] transition starts, the simulation will look for a
-/// [SuperheroVelocity] widget above it and use its velocity to animate the
+/// When a [Heroine] transition starts, the simulation will look for a
+/// [HeroineVelocity] widget above it and use its velocity to animate the
 /// transition.
 ///
 /// If you don't insert one, the velocity will be treated as zero, which is fine
@@ -14,35 +14,33 @@ import 'package:superhero/src/superheroes.dart';
 ///
 /// See the implementation of [DragDismissable] for an example of how to use
 /// this.
-class SuperheroVelocity extends InheritedWidget {
-  /// Creates a new [SuperheroVelocity] widget.
-  const SuperheroVelocity({
+class HeroineVelocity extends InheritedWidget {
+  /// Creates a new [HeroineVelocity] widget.
+  const HeroineVelocity({
     required super.child,
     this.velocity = Velocity.zero,
     super.key,
   });
 
-  /// The current velocity of the next superhero below this widget.
+  /// The current velocity of the next [Heroine] below this widget.
   final Velocity velocity;
 
-  /// Returns the current velocity of the next superhero below this widget.
+  /// Returns the current velocity of the next [Heroine] below this widget.
   ///
   /// If [listen] is true, the listener will be notified when the velocity
   /// changes.
   static Velocity? of(BuildContext context, {bool listen = true}) {
     return listen
         ? context
-                .dependOnInheritedWidgetOfExactType<SuperheroVelocity>()
+                .dependOnInheritedWidgetOfExactType<HeroineVelocity>()
                 ?.velocity ??
             Velocity.zero
-        : context
-                .getInheritedWidgetOfExactType<SuperheroVelocity>()
-                ?.velocity ??
+        : context.getInheritedWidgetOfExactType<HeroineVelocity>()?.velocity ??
             Velocity.zero;
   }
 
   @override
-  bool updateShouldNotify(SuperheroVelocity oldWidget) {
+  bool updateShouldNotify(HeroineVelocity oldWidget) {
     return oldWidget.velocity != velocity;
   }
 }