+++ packages/rivership/lib/rivership.dart
@@ -1,8 +1,8 @@
 /// An opinionated infrastructure package for use with hooks_riverpod.
 library rivership;
 
+export 'package:heroine/heroine.dart';
 export 'package:springster/springster.dart';
-export 'package:superhero/superhero.dart';
 
 export 'src/design/blend_color_tween.dart';
 export 'src/design/simple_widget_states.dart';